research/requirements.txt
anthropic>=0.40.0
pandas>=2.0
httpx>=0.27
pyyaml>=6.0
